LayoutTests/fast/js/mozilla/strict/15.4.4.6-expected.txt

PASS 'use strict'; var a = arr(); [a.pop(), a] threw exception of type TypeError.
PASS var a = arr(); [a.pop(), a] threw exception of type TypeError.
PASS true === true
FAIL 'use strict'; var o = obj(); [Array.prototype.pop.call(o), o] should throw an instance of TypeError
FAIL var o = obj(); [Array.prototype.pop.call(o), o] should throw an instance of TypeError
PASS true === true
 PASSED! 
PASS successfullyParsed is true

TEST COMPLETE
